requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
cryptography
google-generativeai
pydantic
python-dotenv
pandas
